% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_install.R
\name{install.packages}
\alias{install.packages}
\title{Package installation message}
\usage{
install.packages(...)
}
\description{
To protect patient data, internet access is disabled in your Arcus Lab while your data is mounted.
}
