% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_install.R
\name{test_function}
\alias{test_function}
\title{Test function}
\usage{
test_function(...)
}
\description{
To protect patient data, internet access is disabled in your Arcus Lab while your data is mounted.
}
